requirements.txt
Django>=5.1,<5.2
wagtail>=6.4,<6.5
python-dotenv==0.19.1
